% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_id.R
\name{is_ss_id}
\alias{is_ss_id}
\alias{extract_ss_id}
\title{Extract and detect a Smartsheet sheetId}
\usage{
is_ss_id(ss_id)

extract_ss_id(ss_id)
}
\arguments{
\item{ss_id}{A character vector of sheetId(s) or URL(s)}
}
\description{
These functions determine what are SheetIds and extracts sheetIds from URLs.
}
