% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_add_columns.R
\name{ss_add_columns}
\alias{ss_add_columns}
\title{Add columns to an existing sheet}
\usage{
ss_add_columns(ss_id, data, index = 0)
}
\arguments{
\item{ss_id}{The sheetId, permalink, or name of the Smartsheet sheet to read}

\item{data}{A data frame of columns to be added}

\item{index}{The index location where the columns should be added}
}
\value{
A \code{ss_addcolumns_resp} object
}
\description{
Add columns to an existing sheet
}
\examples{
\dontrun{
ss_id = ss_sheetid(ss_write_sheet(paste0("smartsheetr-example-",random_sheet_name())))
ss_add_columns(ss_id, data.frame("FK"=character()), index=1)
ss_read_sheet(ss_id)
# clean up
ss_delete_sheet(ss_id)
}

}
