% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_api.R
\name{ss_api}
\alias{ss_api}
\title{The workhorse function that performs each call to the Smartsheet API}
\usage{
ss_api(FUN, ...)
}
\arguments{
\item{FUN}{An http verb function, typically from the \code{httr} package}

\item{...}{Further parameters passed to the http verb function}
}
\description{
The workhorse function that performs each call to the Smartsheet API
}
