% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_write_columns.R
\name{ss_column_type}
\alias{ss_column_type}
\title{Return the Smartsheet Column Type that aligns with the R class}
\usage{
ss_column_type(r_class)
}
\arguments{
\item{r_class}{A character vector (returned from a call to \code{base::class()})}
}
\value{
A character vector
}
\description{
Return the Smartsheet Column Type that aligns with the R class
}
\details{
See \link{https://smartsheet.redoc.ly/tag/columnsRelated/#section/Column-Types}
}
