% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_sheet.R
\name{ss_column_type_to_class}
\alias{ss_column_type_to_class}
\title{Return an empty vector of the correct class from the smartsheet Column Type}
\usage{
ss_column_type_to_class(ss_column_type)
}
\arguments{
\item{ss_column_type}{A character vector representing a smartsheet Column Type}

\item{r_class}{A character vector (returned from a call to \code{base::class()})}
}
\value{
A character vector
}
\description{
The opposite of \code{\link{ss_column_type}}
}
\details{
See \link{https://smartsheet.redoc.ly/tag/columnsRelated/#section/Column-Types}
}
