% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_write_columns.R
\name{ss_columns}
\alias{ss_columns}
\title{Create a Smartsheets API Columns Object from an R data frame}
\usage{
ss_columns(df)
}
\value{
a data frame
}
\description{
See \href{https://smartsheet.redoc.ly/tag/columnsObjects#section/Column-Object}{Smartsheets API Columns Object Reference}.
This function outputs a data frame that can be used as a \code{columns} argument in \code{ss_write_sheet()}.
}
