% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_write_sheet.R
\name{ss_create_sheet_with_columns}
\alias{ss_create_sheet_with_columns}
\title{Write the initial columns for the a sheet}
\usage{
ss_create_sheet_with_columns(sheet_name, df)
}
\arguments{
\item{sheet_name}{A character vector}

\item{df}{A data frame of columns to be added}
}
\value{
A \code{ss_addcolumns_resp} object
}
\description{
Write the initial columns for the a sheet
}
\details{
See \href{https://smartsheet.redoc.ly/tag/columnsObjects#section/Column-Object}{Smartsheets API Columns Object Reference}.
}
