% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_delete_columns.R
\name{ss_delete_columns}
\alias{ss_delete_columns}
\title{Delete non-primary columns from a given sheet.}
\usage{
ss_delete_columns(ss_id, column_ids = NULL)
}
\arguments{
\item{ss_id}{The sheetId (or permalink) of the table to read}

\item{column_ids}{A vector of the smartsheet rowIds, or NULL to delete all non-primary columns}
}
\value{
A list of ss_resp objects
}
\description{
The primary column(s) cannot be deleted.
}
