% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_delete_rows.R
\name{ss_delete_rows}
\alias{ss_delete_rows}
\title{Delete rows from a given sheet}
\usage{
ss_delete_rows(ss_id, row_ids = NULL)
}
\arguments{
\item{ss_id}{The sheetId (or permalink) of the table to read}

\item{row_ids}{A vector of the smartsheet rowIds, or NULL to delete all}
}
\value{
A list of ss_resp objects
}
\description{
Delete rows from a given sheet
}
