% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_delete_sheet.R
\name{ss_delete_sheet}
\alias{ss_delete_sheet}
\title{Delete a smartsheet}
\usage{
ss_delete_sheet(ss_id)
}
\arguments{
\item{ss_id}{The sheetId (or permalink) of the table to read}
}
\value{
An ss_resp object
}
\description{
Delete a smartsheet
}
