% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_get_sheet.R
\name{ss_get_sheet}
\alias{ss_get_sheet}
\title{Get Sheet Data from a Smartsheet}
\usage{
ss_get_sheet(ss_id)
}
\arguments{
\item{ss_id}{string, The ID of the Smartsheet from which data is to be fetched.}
}
\value{
A \code{tibble} containing the data from the specified Smartsheet.
}
\description{
Gets sheet data from a specified Smartsheet using the Smartsheet API and returns it as a tibble.
}
\examples{
\dontrun{
df = mtcars
ss_id = ss_sheetid(ss_write_sheet(paste0("smartsheetr-example-",random_sheet_name()), data=df))
ss_get_sheet(ss_id)
# clean up
ss_delete_sheet(ss_id)
}
}
