% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_get_sheets.R
\name{ss_get_sheets}
\alias{ss_get_sheets}
\title{Get a data frame describing the Smartsheets available}
\usage{
ss_get_sheets()
}
\description{
Get a data frame describing the Smartsheets available
}
\details{
Note that the environment variable SMARTSHEET_API_TOKEN should be defined in order to run this or any
other \code{smarsheetr} function.
}
