% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_list_sheet_shares.R
\name{ss_list_sheet_shares}
\alias{ss_list_sheet_shares}
\title{List share data for a given sheet}
\usage{
ss_list_sheet_shares(ss_id)
}
\arguments{
\item{ss_id}{The sheetId, permalink, or name of the Smartsheet sheet to read}
}
\value{
A dataframe
}
\description{
List share data for a given sheet
}
\examples{
ss_id = ss_sheetid(ss_write_sheet(paste0("smartsheetr-example-",random_sheet_name())))
ss_list_sheet_shares(ss_id)
# clean up
ss_delete_sheet(ss_id)

}
