% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_replace_sheet.R
\name{ss_replace_sheet}
\alias{ss_replace_sheet}
\title{Replace the contents of a sheet with a new data frame}
\usage{
ss_replace_sheet(ss_id, df)
}
\arguments{
\item{ss_id}{The sheetId (or permalink) of the table to read}

\item{df}{The new sheet contents}
}
\value{
A named list of ss_resp objects
}
\description{
Replace the contents of a sheet with a new data frame
}
