% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_api.R
\name{ss_resp_data_to_dataframe}
\alias{ss_resp_data_to_dataframe}
\title{Helper to rbind lists in a list into a data frame}
\usage{
ss_resp_data_to_dataframe(resp_data)
}
\arguments{
\item{resp_data}{A list of lists}
}
\description{
Helper to rbind lists in a list into a data frame
}
