% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_row_ids.R
\name{ss_row_ids}
\alias{ss_row_ids}
\title{List row ids for a given sheet}
\usage{
ss_row_ids(ss_id)
}
\arguments{
\item{ss_id}{The sheetId, permalink, or name of the Smartsheet sheet to read}
}
\value{
A numeric vector
}
\description{
Returns a vector of the Smartsheet internal row ids for a given sheet
}
\examples{
\dontrun{
df = data.frame(PK=c(1,2), FK=c("a","b"))
ss_id = ss_sheetid(ss_write_sheet(paste0("smartsheetr-example-",random_sheet_name()), data=df))
ss_row_ids(ss_id)
# clean up
ss_delete_sheet(ss_id)
}

}
