% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_sheetid.R
\name{ss_sheetid}
\alias{ss_sheetid}
\title{Get a smartsheet sheetId from a response}
\usage{
ss_sheetid(resp)
}
\arguments{
\item{resp}{A ss_resp object}
}
\value{
A numeric sheetId
}
\description{
Get a smartsheet sheetId from a response
}
