% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{to_json}
\alias{to_json}
\title{Convert list to a JSON character vector}
\usage{
to_json(...)
}
\arguments{
\item{...}{Passed on to \link[jsonlite]{toJSON}}
}
\description{
The default method of jsonlite::toJSON() returns values wrapped as vectors, e.g. \verb{\{"x":[1]\}}. The
smartsheets API complains about this, so we must remove it by using the auto_unbox parameter.
}
\keyword{internal}
