% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_sheet.R
\name{unlist_and_replace_null}
\alias{unlist_and_replace_null}
\title{Helper function to replace NULL values with NA, and unlist, which is useful in converting nested lists
to data frames}
\usage{
unlist_and_replace_null(l)
}
\arguments{
\item{l}{A list}
}
\description{
Helper function to replace NULL values with NA, and unlist, which is useful in converting nested lists
to data frames
}
