% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{url_query}
\alias{url_query}
\title{Build a URL query}
\usage{
url_query(l)
}
\arguments{
\item{l}{A names list of query parameters}
}
\description{
Build a URL query
}
\examples{
url_query(list(pageNumber=1, pageSize=100))

}
