% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ss_id.R
\name{validate_ss_id}
\alias{validate_ss_id}
\title{Validate or get the sheetID from a numeric/character vector}
\usage{
validate_ss_id(ss_id)
}
\arguments{
\item{ss_id}{A smartsheet sheet name, permalink, of sheetId}
}
\value{
A smartsheets sheetId
}
\description{
This function validates a single ss_id is passed in and returns a smartsheets sheetId
}
