% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_add_columns.R
\name{ss_add_columns}
\alias{ss_add_columns}
\title{Add columns to an existing sheet}
\usage{
ss_add_columns(ss_id, df, index = 0)
}
\arguments{
\item{ss_id}{The sheetId (or permalink) of the table to read}

\item{df}{A data frame of columns to be added}

\item{index}{The index location where the columns should be added}
}
\value{
A \code{ss_createsheet_resp} object
}
\description{
Add columns to an existing sheet
}
\details{
See \href{https://smartsheet.redoc.ly/tag/columnsObjects#section/Column-Object}{Smartsheets API Columns Object Reference}.
}
