% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_add_rows.R
\name{ss_add_rows}
\alias{ss_add_rows}
\title{Add rows to a sheet.}
\usage{
ss_add_rows(ss_id, df, column_ids = NULL)
}
\arguments{
\item{ss_id}{The sheetId (or permalink) of the table to read}

\item{df}{A data frame of rows to add}

\item{column_ids}{A vector of the columnIds of the smartsheets sheetId. If \code{NULL}, this will be obtained.}
}
\description{
Add rows to a sheet.
}
