% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_sheet.R
\name{ss_cols_to_dataframe}
\alias{ss_cols_to_dataframe}
\title{Helper function to take columns data and create a data frame.}
\usage{
ss_cols_to_dataframe(ss_cols_data)
}
\arguments{
\item{ss_cols_data}{A data frame}
}
\description{
Helper function to take columns data and create a data frame.
}
