% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_get.R
\name{ss_get}
\alias{ss_get}
\alias{ss_response_parse}
\title{Smartsheet GET}
\usage{
ss_get(path)

ss_response_parse(resp)
}
\arguments{
\item{path}{A character vector to add to the API url. See (https://smartsheet.redoc.ly/#section/Introduction)
for more information.}

\item{resp}{An httr::response object}
}
\value{
An httr::response object
}
\description{
This is a simple wrapper around the GET function in the \href{https://smartsheet.redoc.ly/}{Smartsheet API 2.0}
}
\details{
Note that the environment variable SMARTSHEET_API_TOKEN should be defined in order to run this or any
other \code{smarsheetr} function.
}
\section{Functions}{
\itemize{
\item \code{ss_response_parse()}: Parse the response of an ss_get() call.

}}
