% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_api.R
\name{ss_get}
\alias{ss_get}
\alias{ss_post}
\title{Smartsheet GET}
\usage{
ss_get(path, ...)

ss_post(path, body, ...)
}
\arguments{
\item{path}{A character vector to add to the API url. See (https://smartsheet.redoc.ly/#section/Introduction)
for more information.}

\item{body}{A list of objects}
}
\value{
An httr::response object
}
\description{
This is a simple wrapper around the GET function in the \href{https://smartsheet.redoc.ly/}{Smartsheet API 2.0}
}
\details{
Note that the environment variable SMARTSHEET_API_TOKEN should be defined in order to run this or any
other \code{smarsheetr} functions.
}
\section{Functions}{
\itemize{
\item \code{ss_post()}: SmartSheet POST

}}
