% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_list_sheets.R
\name{ss_list_sheets}
\alias{ss_list_sheets}
\title{Get a data frame describing the smartsheets available}
\usage{
ss_list_sheets()
}
\description{
Get a data frame describing the smartsheets available
}
\details{
Note that the environment variable SMARTSHEET_API_TOKEN should be defined in order to run this or any
other \code{smarsheetr} functions.
}
