% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_sheet.R
\name{ss_read_sheet}
\alias{ss_read_sheet}
\title{Reads a Smartsheet sheet into an R data frame}
\usage{
ss_read_sheet(ss_id)
}
\arguments{
\item{ss_id}{The sheetId, permalink, or name of the Smartsheet sheet to read}
}
\value{
A tibble::tbl_df object
}
\description{
Reads a Smartsheet sheet into an R data frame
}
\examples{
\dontrun{
df = mtcars
ss_id = ss_sheetid(ss_write_sheet(paste0("smartsheetr-example-",random_sheet_name()), data=df))
ss_read_sheet(ss_id)
# clean up
ss_delete_sheet(ss_id)
}

}
