% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_read_sheet.R
\name{ss_read_sheet}
\alias{ss_read_sheet}
\title{Require either a sheet name or a link,}
\usage{
ss_read_sheet(ss_id)
}
\arguments{
\item{ss_id}{The sheetId (or permalink) of the table to read}
}
\value{
A tibble::tbl_df object
}
\description{
Require either a sheet name or a link,
}
