% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_replace_sheet.R
\name{ss_replace_sheet}
\alias{ss_replace_sheet}
\title{Replace the contents of a sheet with a new data frame}
\usage{
ss_replace_sheet(ss_id, data)
}
\arguments{
\item{ss_id}{The sheetId, permalink, or name of the Smartsheet sheet to read}

\item{data}{A data frame}
}
\value{
A named list of ss_resp objects
}
\description{
Replace the contents of a sheet with a new data frame
}
\examples{
\dontrun{
ss_id = ss_sheetid(ss_write_sheet(paste0("smartsheetr-example-",random_sheet_name())))
ss_replace_sheet(ss_id, data=mtcars)
ss_read_sheet(ss_id)
# clean up
ss_delete_sheet(ss_id)
}

}
