% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_write_sheet.R
\name{ss_write_sheet}
\alias{ss_write_sheet}
\title{Create a sheet}
\usage{
ss_write_sheet(sheet_name, df = data.frame(PK = character()))
}
\arguments{
\item{sheet_name}{A character vector}

\item{df}{A data frame}
}
\value{
A smartsheetsr response object
}
\description{
Creating a sheet requires either a template or a set of columns
(see https://smartsheet.redoc.ly/tag/sheets#operation/create-sheet-in-sheets-folder). This function only
allows for the columns option.
}
